+++ source/utils/h2_numeric.hpp
@@ -1,4 +1,4 @@
 struct h2_numeric {
    static bool is_bin_string(const char* s);
-   static const char* sequence_number(size_t sequence, int shift = 1);
+   static const char* sequence_number(size_t sequence, size_t shift = 1);
 };
